feat(EssValue API) : ess value 조회 API
-충전 끝나면 ess value를 차감한다.
-ess value를 조회하는 api 구현

+++ src/main/java/com/example/quaterback/api/domain/station/domain/ChargingStationDomain.java
@@ -21,7 +21,7 @@ public class ChargingStationDomain {
     private String address;
     private LocalDateTime updateStatusTimeStamp;
     private StationStatus stationStatus;
-    private Integer essValue;
+    private Double essValue;
 
     public void updateStationStatus(StationStatus status) {
         if (!this.stationStatus.equals(status)) {
@@ -30,7 +30,7 @@ public class ChargingStationDomain {
         }
     }
 
-    public void updateStationEssValue(Integer essValue){
+    public void updateStationEssValue(Double essValue){
         this.essValue = essValue;
     }
